changed x and y position to double to allow for more accurate movement speeds. also you can change the way the wavesetup class handels how many attackers to send out per wave

+++ PicnicProtector/src/Attacker.java
@@ -1,13 +1,13 @@
 public abstract class Attacker {
 
 	//movementspeed is how many pixels per update attacker should move
-	protected int x, y, movementSpeed;
+	protected double x, y, movementSpeed;
 	
 	//will be set depending on the map
 	protected String type, moving="down";
 	
 
-	public Attacker(int x, int y){
+	public Attacker(double x, double y){
 		
 		this.x = x;
 		this.y = y;
@@ -53,14 +53,14 @@ public abstract class Attacker {
 	}
 	
 	public int getX(){
-		return x;
+		return (int) x;
 	}
 	
 	public int getY(){
-		return y;
+		return (int) y;
 	}
 	
-	public void repos(int x, int y){
+	public void repos(double x, double y){
 		this.x = x;
 		this.y = y;
 	}
